Scene: relocate non-scene helpers out of the scene module
Move node-by-path navigation onto `RenderNode` (`node_at_path`/`node_at_path_mut`)
and `transformed_rect_extents` into a new `geometry` module; inline the
`apply_transform` and `collect_layout_children` pass-throughs into their callers.
The scene module is left holding only the stacking-context model.
