+++ takumi-core/src/lib.rs
@@ -12,6 +12,7 @@ pub mod shadow;
 pub mod text_processing;
 
 pub mod error;
+pub mod geometry;
 pub use takumi_css::keyframes;
 pub mod resources;
 pub mod scene;
